Refactor: ChannelType 위치 변경 및 관련 참조 수정 - `entity.entityType`에서 `entity.role`로 클래스 위치 변경 - 기존 클래스 참조 위치 모두 수정
