+++ src/main/java/com/sprint/mission/discodeit/dto/channelDto/ChannelDto.java
@@ -3,7 +3,7 @@ package com.sprint.mission.discodeit.dto.channelDto;
 import com.sprint.mission.discodeit.dto.userDto.UserDto;
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.Message;
-import com.sprint.mission.discodeit.entity.entityType.ChannelType;
+import com.sprint.mission.discodeit.entity.role.ChannelType;
 import lombok.Builder;
 
 import java.time.Instant;
