+++ src/main/java/com/sprint/mission/discodeit/entity/role/ChannelType.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity.entityType;
+package com.sprint.mission.discodeit.entity.role;
 
 import lombok.Getter;
